voicebot-python/requirements.txt
azure-ai-language-conversations==1.1.0
azure-ai-translation-text==1.0.0b1
azure-cognitiveservices-speech==1.33.0
azure-core==1.29.5
certifi==2023.11.17
charset-normalizer==3.3.2
dnspython==2.4.2
idna==3.4
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
pydub==0.25.1
pymongo==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.8.0
urllib3==2.1.0
